CleanUpGithub
Hi, I’m part of #CleanUpGithub, an event trying to refactor equality in the technology industry. This repo appears to serve no purpose other than to offend. Repos like this act as a barrier for entry to programming for a significant number of people - and we all want to see more coders! This message is from a human, not a bot.